MNT: status-related stuff to its own module
